Extract interface to types.ts
